edgepaint furtherest_point: remove disabled debugging code
This code was (1) assuming /tmp exists and is writable and (2) did not close the
file it opened. It is also not possible to reach this code, because `Verbose`
cannot be set above 10 without editing the source code. If debugging the
quadtree structure is desirable, we should find a cleaner way to expose this.