tclpkg: fix discovery of package parts on non-Linux
3fcc3c62e7d8fda22ca021e2317bd2cb567b8113 adapted this script to recognize shared
objects in addition to libtool libraries, but it failed to account for shared
objects having a different suffix than “.so” on some platforms. For example, on
macOS the “.dylib” suffix of tclpathplan.dylib would go unrecognized and the
script would produce incorrect results.

This alternative change only tries to derive a base when the input looks like a
libtool library. In other cases it just takes the library as-is, supporting any
prefix/suffix the current platform uses.

Gitlab: #2530