CMake: support for building gv_tcl
It is not clear to me why tclpkg/gv/Makefile.am chooses to use “0” as the
version number in the call to mkpkgindex.sh where the other TCL libraries use
`$(VERSION)`, but this change replicates it in the CMake build system anyway.

The man page for gv_tcl is generated, in contrast to the other man pages handled
by the CMake build system. So the rules introduced for this one anticipate
finding gv.3tcl in the build directory instead of the source directory.

The script that generates the man pages produces them for all the language
bindings at once. The rules added in this commit to describe this leads to the
odd situation where the CMake build system now generates (but does not install)
man pages for language bindings it does not yet build. But it seemed better to
be “honest” with the build system about the inputs and outputs of this step
upfront.

The generated gvTCL_wrap.cxx file needs to be excluded from coverage testing as
with other generated files. But for reasons unclear to me excluding it in the
normal way passes the init run but fails during analysis:

  geninfo: ERROR: 'exclude' pattern
    '/builds/graphviz/graphviz/build/tclpkg/gv/gv_tcl.cpp/gvTCL_wrap.cxx' is
    unused.

Gitlab: #2530