xlib plugin: set close-on-exec on inotify file descriptor
Because this code is within a plugin, it does not know what the application that
has loaded it may be doing. In particular, the application may be running
`fork`+`exec` concurrently with the plugin code using inotify. The child process
created by this would inadvertently inherit a copy of the inotify file
descriptor, something the plugin code definitely does not intend. In fact, this
plugin itself `fork`+`exec`s to run a web browser. It seems clear the original
author did not intend the web browser to inherit a copy of its inotify file
descriptor.

This change splits the existing `inotify_init` call into three paths:

  1. `inotify_init1(IN_CLOEXEC)`, as implemented by Linux ≥2.6.27, Glibc ≥2.9.
  2. `inotify_init1(IN_CLOSEXEC)`, as implemented by libnotify on FreeBSD.
     Unfortunately this seems to have been implemented with a different name for
     the close-on-exec flag.
  3. `inotify_init()`+`fcntl` for Linux <2.6.27. In this case, there is a race
     window between creating the file descriptor and setting close-on-exec, but
     this seems unavoidable.