CMake: fix static linking gvpack against neato layout plugin
When building static libraries (`-DBUILD_SHARED_LIBS=OFF`), linking of gvpack
failed because the neato layout plugin had been built as a shared library. My
understanding is that in this scenario the linker is looking for an unmangled
implementation but instead finds only mangled symbols.¹

Making the plugin buildable as a static library does not seem like the right fix
because, by definition, the plugins are intended to be dynamically loaded. This
change instead “unwraps” the link command into using the plugin’s object files.
We need to then explicitly obtain and use its dependencies because we can no
longer rely on CMake’s target analysis to do this for us.

This change chooses to use a conditional warning to avoid dropping compatibility
with older CMake. Once environments with CMake ≤3.21 drop out of support, we
should remove this warning and just bump up the required CMake version in the
top level file.

Gitlab: !3702
Reported-by: Daniel Nicoletti <daniel.nicoletti@kdab.com>

¹ https://stackoverflow.com/a/32535257