CMake: make '-DBUILD_SHARED_LIBS=OFF' build plugins statically and not install
The plugins were being forced to be built as shared libraries because this is
how they are anticipated to be loaded. However if you attempt a static build on
Windows, linking `gvpack` fails because it tries to link against a
built gvplugin_neato_layout archive but this plugin has been built as a shared
library. There are various ways to work around this, but when then trying to
build a static `dot` binary these work arounds need to also be applied there.
The end result of this is some statically linked binaries accompanied by some
shared libraries they do not use.

A simpler approach is to accept users’ requirements are probably a shared xor
static build. That is, users likely want to build (1) `dot` and a collection of
plugins as shared libraries or (2) a static monolithic `dot` with plugins
compiled in; never some combination of (1) and (2) at once. If they want both
(1) and (2) they can run two separate CMake compilations.

This change does not yet allow a static `dot` to be built, but it sets up the
plugins such that this can be done in future. Installation is suppressed because
the plugins are not easily usable as archives outside of the build tree; while
we could install these .a files, they come with no accompanying headers and can
only be linked against by redeclaring their symbols in application code. Such
redeclarations are inherently fragile (they lead to silently broken code if the
actual plugins’ implementations ever change), so we do not want to encourage
this kind of practice.

Gitlab: !3702
Suggested-by: Daniel Nicoletti <daniel.nicoletti@kdab.com>