CMake: make '-DBUILD_SHARED_LIBS=OFF' define 'dot' with deps statically linked
This replicates something similar to `dot_static` in the Autotools build system.
Rather than allow both shared and static binaries to be built at once, we ask
the CMake user to choose between one or the other. This seems to fit with the
CMake model with less friction. If users want both, they can run two separate
compilations.

Gitlab: !3702
Reported-by: Daniel Nicoletti <daniel.nicoletti@kdab.com>