Revert "again try to stop it picking up old libgvc from someplace!"
This reverts commit 38fffd80b95eeb0a395c6525815b778f37953b80. While it is clear
what the aim of this commit was (avoiding linking against a system-provided
libgvc), it is unclear how it makes progress towards that goal. As the Autotools
docs¹ explain, libgvc was already in the list of dependencies here:

  If `_DEPENDENCIES` is not supplied, it is computed by Automake. The
  automatically-assigned value is the contents of `_LDADD` or `_LIBADD`, with
  most configure substitutions, `-l`, `-L`, `-dlopen` and `-dlpreopen` options
  removed.

So this change, if anything, makes the build ordering _weaker_ than it was
before.

¹ https://www.gnu.org/software/automake/manual/html_node/Program-and-Library-Variables.html