CMake: use 'UseSWIG' v2; remove incorrect 'OUTFILE_DIR' setting for TCL bindings
The `OUTFILE_DIR` option controls not the output file itself, but the directory
in which it is placed. So this was leading to `swig` being invoked with
something like `-o gv_tcl.cpp/gvTCL_wrap.cxx`. Nothing in the build systems seem
to depend on these paths (the Autotools build system seems to just use these to
avoid bindings for different languages overwriting one another), so lets use a
more modern way of interacting with SWIG that namespaces the outputs
per-language.

Gitlab: #2530