CMake: add C# bindings support
Unlike the corresponding Autotools support, there is no discovery of the C#
compiler or the runtime’s includes/libraries. The compiler seems only necessary
for testing. The generated bindings use only Graphviz headers and the C++
standard library and thus do not need the C# runtime present in the build
environment.

This change has the counterintuitive behavior wherein the C# binding’s man page,
gv.3sharp, is only installed if `-Denable_tcl=ON` is also set. Given the man
page is generated by a TCL script, I do not know how to avoid this situation.

Gitlab: #2530