gvedit: Efficient QString usage
Added definitions -D that make sure
no implict conversion from ASCII is done,
this is so the QString usage is done knowing
what is being converted and which is the codec
in use.

Creating a QString{"foo"} without those definitions
will incur on the slow UTF-8 conversion, while
the text is known to be latin-1.

Because of Qt5.9 in CI it's not possible to use
QStringView, which reduces allocations, and
is implicit called when using the u"" prefix.

Some places were only needing a QChar, so u''
implict creates one.

QStringLiteral preallocates the static strings
thus not requiring allocation or codec conversion.

QLatin1String allows to "mark" string that are
known to be latin1.