CMake: add pkgIndex.tcl support
This implements the equivalent of the `pkgIndex.tcl` rule in the Autotools’
build system in tclpkg/Makefile.am.

There are ways to work around the lack of `cmake -E cat` on CMake < 3.18. But
this does not seem worth it. When Ubuntu 20.04 drops out of support, we can
raise the minimum required CMake version.

Gitlab: #2530