glcomp glCompSetAddNewTexLabel: fix use of uninitialized variables
When running the CMake build in release mode, GCC 11.4.0 notices:

  lib/glcomp/glcomptexture.c: In function ‘glCompSetAddNewTexLabel.part.0’:
  lib/glcomp/glcomptexture.c:128:9: warning: ‘width’ may be used uninitialized
    [-Wmaybe-uninitialized]
    128 |     t = glCompSetAddNewTexture(s, width, height, data, is2D);
        |         ^
  lib/glcomp/glcomptexture.c:102:13: note: ‘width’ was declared here
    102 |     int Er, width, height;
        |             ^
  lib/glcomp/glcomptexture.c:128:9: warning: ‘height’ may be used uninitialized
    [-Wmaybe-uninitialized]
    128 |     t = glCompSetAddNewTexture(s, width, height, data, is2D);
        |         ^
  lib/glcomp/glcomptexture.c:102:20: note: ‘height’ was declared here
    102 |     int Er, width, height;
        |                    ^

This function appears incorrect in other ways (e.g. `surface` can statically be
determined to be `NULL` throughout this function, yet it is written in a style
as if the author believed this parameter was passed by reference to
`glCompCreatePangoTexture`), but this at least fixes an instance of Undefined
Behavior.