expr: remove unnecessary stringizing macro
This code was written using two levels of macros as a common trick for expanding
an argument you believe itself to be a macro before stringizing it.¹ It is not
clear `intmax_t` and `uintmax_t` are in scope here to be expanded because the
file does not #include <stdint.h>, but moreover these types are not macros in
contemporary C99 implementations anyway. When
1c2bab8289f901c9981eca63f426291ade9b1836 altered the types in use here, it made
the use of this macro a no-op.

¹ https://gcc.gnu.org/onlinedocs/cpp/Stringizing.html