CMake: fix rpaths in plugins and language bindings
489f56c3deacb652daf37da3000def20c5184881 set the rpaths for all components,
including the plugins and language bindings.¹ However, the plugins and language
bindings live deeper in the install hierarchy than the binaries and libraries.
So “$ORIGIN/..” (or on macOS, “@executable_path/..”) in their context does not
name the install root.

This change corrects the problem by setting rpath on a more fine-grained basis.
`list(APPEND …)` appears scoped to the current directory and below, so this is
sufficient to set it correctly for all installed artifacts.

It is relatively unlikely that a user would try to load a Graphviz plugin
without also intending to load libcgraph or libgvc (the only scenario in which
this comes into play), but loading a Graphviz language binding without
explicitly loading libcgraph is an expected use case.

Gitlab: fixes #2567

¹ Language bindings were not actually integrated into the CMake build system at
  the time of 489f56c3deacb652daf37da3000def20c5184881, so technically this
  aspect of the change was latent until
  f21c617748d698fd8f711a145edb0038dd8c4534.