CMake: split discovery of PangoCairo for native Windows and MinGW
This makes it clear which paths are valid for each platform and
facilitates making platform specific changes in upcoming commits.

This is currently a lot of duplication, but upcoming commits that will
bring pkgconf into use also on these platforms will reduce this
substantially. The native Windows PangoCairo library must however first
be updated to a version that supports pkgconf.

GitLab: #2557