CMake: split discovery of GTS for native Windows and MinGW
This makes it clear which paths are valid for each platform and
facilitates making platform specific changes in upcoming commits.

This is currently a lot of duplication, but upcoming commits that will
bring pkgconf into use also on these platforms, will reduce this
substantially. A GTS library that supports pkgconf must however first be
installed for native Windows. We currently don't have any version of GTS
installed in `windows/dependencies/libraries`, but the version that is
possible to discover with the current dicovery mechanism presumably
don't support pkgconf. We will therefore keep the current discovery
until we upgrade.

GitLab: #2557