gvc: remove rounding up to page alignment during write allocation
This kind of thing is no longer relevant on contemporary operating systems.
Firstly, 4K is no longer the universal page size. Secondly, libc allocators have
a much more sophisticated and accurate understanding of optimal allocation
requests and are able to do a better job of this themselves if we are simply
more honest with our requirements. This change squashes 4 -Wsign-conversion
warnings.