neatogen: fix shorter adjust modes hiding longer ones
Due to the way `adjustMode` looked at the prefix of attribute values to match
and the ordering of the adjust modes to put prefixes before longer values (e.g.
`"scale"` appears before `"scalexy"`), it was not possible to select the modes
`"scalexy"`, `"ortho_yx"`, `"orthoxy"`, `"orthoyx", `"portho_yx"`, `"porthoxy"`,
and `"porthoyx"`. This bug was introduced in
55ff80950bdbb41f030fa7b07b1d006945182057.

Gitlab: fixes #2563
Reported-by: Giorgi Kitsmarishvili