CI: upgrade Python 3.6 → 3.11 on Rocky 8
The Rocky Linux 8 environment was using the system Python (package
`platform-python`) that is stuck on Python 3.6. Using the `python` package will
allow us to move beyond being pinned to Python 3.6 which is already EOLed.

Because this was the last officially supported platform using < Python 3.8, we
can now require a baseline of Python 3.8, the lowest Python version not yet
EOLed.

Gitlab: closes #2549