refactor round_corners(): split symbol_draw() from round_corners()
round_corners() is bloated and extremely long
and needs to be refactored.

Split symbol_draw() from round_corners() with
minimal change and intentionally without reordering the code
at price of introduction a forward declaration.

Ideally, it is preferable to keep the number of forward declarations
to a minimum.

Further plan is to make it take a bounding box as input
instead of array of verities to reduce complex computations.