gvloadimage_webp: remove legacy Windows inline linking pragmas
The build system already handles the link libraries.

The immediate reason for doing this now is that an upcoming commit
uppdates the GD library and its WebP library dependency. This enables
building the webp plugin, exposing the outdated libraris in the removed
pragmas. Without this, the following error would occur:

LINK : fatal error LNK1104: cannot open file 'libcairo-2.lib' [C:\Users\magjac\graphviz\build\plugin\webp\gvplugin_webp.vcxproj]

The correct name of the Cairo library in contemporary versions is
'cairo.lib'.

GitLab: #1786, #1788