use 'pkg-config' where possible to locate link libraries during testing
This provides an extra safeguard against problems like that fixed in the prior
commits. This is not foolproof – exporting `LIBRARY_PATH` as described in
DEVELOPERS.md easily masks problems with the .pc files because malformed `-L…`
content from the .pc files causes no harm and the `LIBRARY_PATH` export makes
the Graphviz libraries findable despite this – but it is better than the prior
situation.

Gitlab: #2474, !3504, !3594, !3875, #2571