cgraph generic list: unconditionally call a destructor
This removes a minor foot gun. Because of the casts used to access the
destructor, it was possible to pass in a completely mismatched function as the
destructor (e.g. `DEFINE_LIST_WITH_DTOR(ints, int, printf)`). This would
generate some warnings, but would still compile, and then crash at runtime. This
rephrasing forces such scenarios to fail at compile time.