neatogen: fix divergent definition of what indicates a cluster
ae160d36572253aa4edfdb14fefe2729bbdfc771 added an alternative to prefixing the
name of a (sub)graph with “cluster” to indicate it is a cluster; a
`cluster=true` attribute. Unfortunately various parts of the code base have hand
rolled their own inline logic for checking whether a graph is a cluster, and
these other instances were not updated at the same time. These discrepancies
mean Graphviz libraries can disagree about whether a graph is a cluster, causing
misbehavior and crashes.

Similar to 4e59d8f3a74f32f6af1dc2321f0ce6f66af16908, this change fixes one of
these discrepancies by replacing it with the libcommon implementation.

Note that there remains one instance of `startswith(…, "cluster")` in
neatoinit.c. This instance appears to be checking the name of a _node_, not a
graph. So presumably should not be updated in this change.

Gitlab: #2187