Autotools: remove unused 'pango_font_map_create_context' discovery
e01ee1fb8b92efdfb208e1a175972d44748de759 removed the last use of this, seemingly
mistakenly overlooking the discovery of this function performed here. This API
arrived in Pango 1.22 and the lowest Pango version on any currently supported
platform is 1.42.4 on Rocky Linux 8. So presumably it is fine to unconditionally
assume the existence of it now.