pango plugin pango_textlayout: remove 1024-character limit on constructed font
There seems no reason the information being printed to this buffer is guaranteed
to be less than 1024 characters. This most likely fixes a latent buffer
overflow. This introduces a memory leak (the memory backing the static `buf` is
never freed) but there seems no easy way to avoid this with the semantics of the
current API. I.e. that the caller of `gvtextlayout_engine_t.textlayout` expects
not to have to free the returned `fontpath` pointer.