CI: correct case of CMake variables in windows build script
Commits e4ca1b769, 85d3989f1, 28349cfac and a7378639b changed the CMake
variables `ENABLE_LTDL`, `WITH_EXPAT`, `WITH_GVEDIT` and `WITH_ZLIB` to
upper case, but failed to change these variables in the Windows build
script.

There was no impact on our CI builds because of this, since these
variables default to `AUTO` and the CMake discovery causes them to be
set to the same values in CI as this script explicitly sets them to.