Revert "Revert "API BREAK: remove now unused 'BF2B'""
This reverts commit 102c62ab1adc400ba5d1b02235ae36cb581a8498. Following the
previous change, we can once again remove `BF2B`.

This change takes the unorthodox step of _removing_ a changelog entry and
stepping the upcoming version _backwards_. The entry being removed was for the
revert in 102c62ab1adc400ba5d1b02235ae36cb581a8498 and there has been no release
since then. Following the present commit, there is no API break and users do not
need to be exposed to this back-and-forth in the upcoming release.

Gitlab: #2572