pack: uniformly compute on floating-point representations
This is a reattempt of 83806f043bdcd312c032e3b2f7eb1cd00e3918ab which was
subsequently reverted in 48c8f9c5308e7613870f9262af6124f1fcc25a26. The intent of
this change is to remove assumptions that rounded `double` values fit in an
`int`. A beneficial side effect is the removal of 32 compiler warnings.

In contrast to 83806f043bdcd312c032e3b2f7eb1cd00e3918ab, this change takes the
approach of rounding `double` values everywhere the previous code was implicitly
rounding. No attempt was made to examine what subset of these locations are
actually necessary for correctness of the algorithm, but rather every one of
these was conservatively preserved.

This change was validated by confirming `test_2572` still passes, as well as
manually inspecting the fdp and sfdp results on this test case and confirming
they are reasonable.

The rounding sometimes incurs repetitive code. E.g. calls to the `CELL` macro
followed by a round. No attempt was made to de-dupe or outline this. This can
probably be improved in a future change.

This change takes the unorthodox step of _removing_ a changelog entry. The entry
being removed was for the revert in 48c8f9c5308e7613870f9262af6124f1fcc25a26 and
there has been no release since then. So after the present commit, there is no
need for users to see this changelog entry.

Gitlab: #2572