core plugin: fix printed TCL tags to be usable pointers
The TK output format was using IDs to compose a tag for use in output. However
the TCL bindings expect tags to contain a usable pointer they can then use to
locate the corresponding entity. This meant creating a graph in a TCL script or
interactive session would give you handles to the created nodes and edges you
then could not use.

This appears to have been broken for a very long time, and possibly by
contributing factors external to Graphviz. In particular, the general transition
from 32-bit to 64-bit hardware appears to have exposed some assumptions within
Graphviz that `sizeof(int) == sizeof(void *) == 4`. So there have been points in
history where the pointer component of a tag was printed as a 32-bit number,
making it usable on 32-bit platforms but not on 64-bit platforms.

When exactly these changes happened and which platforms were affected is
difficult to conclusively determine. E.g.
fafbc9b66f81069a048265f4e9a0a76ca7195c70 appears to have had a positive effect
on the situation, but it is within the range
[07d27fbab1dece9e52fa28b32a3643c967e2d262,
89f6e0d6891f9fbb3258f2cd2165c7839ee94e27) during which `require Tcldot` and the
TCL bindings as a whole were unusable.

Gitlab: #208, #1564, #2370, fixes #2568
Reported-by: Felix
Suggested-by: Felix