CI: merge ASan build and test jobs and stop archiving the build artifacts
This avoids overwriting the normal Ubuntu 24.04 build artifacts which
are intended to be released and not ASan instrumented.

This also incorporates parts of the ctest job templates into the merged
job and removes those now unused templates.

This change also results in test programs built by the test suite will
be compiled with `-Werror` and `-Wno-error=deprecated-declarations`.

Fixes #2574.