CI: activate Python virtual environment before anything else in deb build template
The immediate effect of this is that Python scripts run by the build
system, e.g. Graphviz version generation, will run with the virtual
environment. This currently has no observable consequences, but ensures
that anything added in the future will run with the virtual environment.

This ensures consistency across builds when an upcoming commit that
merges the ASan build and test jobs without using this template makes
the same change.

GitLab: #2574