tclpkg: port man page generator to Python
A combination of TCL scripts was used to generate man pages for the
SWIG-generated language bindings. A problem with this approach is that it
requires the TCL interpreter installed (an increasingly uncommon default) even
if no TCL components are being built. The CMake build system only used TCL
conditionally, which led to scenarios where you could build the bindings but not
their documentation if you did not have the TCL interpreter.

This change ports this generation to Python. In contemporary environments,
Python is more common and more widely known. As a result, the TCL interpreter is
no longer required by either the Autotools or CMake build system for these
steps. In porting, this change also introduces some comments and converts some
strings to first class types.

Within the Autotools build system, this generation is done at `make dist` time.
So consumers of the portable source tarball still do not need Python. In the
CMake build system, Python was already required so this does not introduce an
additional dependency.