stop distributing Visual Studio build files in the portable source tarball
Since fa81bad7317749efd750e76b495d178d9eb55c78, which was released in
12.0.0, we no longer distribute Windows binaries built with Visual
Studio. This is the next step in removing support for Visual Studio and
only support CMake for Windows builds.

Since https://gitlab.com/graphviz/graphviz/-/merge_requests/3862,
https://gitlab.com/graphviz/graphviz/-/merge_requests/3897 and
https://gitlab.com/graphviz/graphviz/-/merge_requests/3911 the Windows
CMake builds are on par or better than the MSBuild builds. Possibly with
the exception of Smyrna which is built by the MSBuild builds, but since
we do not distribute the necessary DLLs to run it for reasons described
in https://gitlab.com/graphviz/graphviz/-/issues/1829#note_2010004585,
it's not directly usable.

GitLab: related to https://gitlab.com/graphviz/graphviz/-/issues/2530.