CMake: add JavaScript bindings support
This change hardcodes discovery for libjavascriptcore 4.0 and 4.1, the only
versions of this library available in supported operating systems and supporting
the API that SWIG expects.

Note there is no man page for these bindings.

Unlike the Autotools build, this does no discovery of the JavaScript Core
compiler, `jsc`. This is not required for building the bindings.

Gitlab: #2530