CMake: add Go bindings support
Unlike the Autotools build, this auto-detects `INTGOSIZE` based on the target’s
word size. This seems simpler and more likely correct. It works when
cross-compiling too.

This implementation installs an extra file, gv_go.so, that is not installed by
the Autotools implementation. I do not know exactly why this is, but it seems to
be intentional in the implementation of `swig_add_library`.

Gitlab: #2530