circogen appendNodelist: take index at which to insert instead of one before it
As a remnant of the old linked-list implementation of node lists, this function
was accepting an index of one before where to insert the new node. This came
about because of the straightforward way this was ported to the generic list in
bd35c9669be37edf55ee89055241aa89f0c95325. Using the actual index to insert at
leads to more intuitive code and will allow future simplifications.