fix gvpr’s handling of empty strings as 'printf' parameters
In transitioning from sfio’s I/O to libc stdio,
8ce6b211ce7b9ebad3e30fa909effebf59955a93 introduced a bug wherein printing
zero-length strings would be misinterpreted as an I/O error. This was mostly
ignored by libexpr and libgvpr and not visible to the user, except that it
terminated the current `[s]printf` call. So if there was anything following the
empty string, this would be truncated. E.g. `printf("hello%s world\n", "")` used
in one of the test cases fixed here would previously print just “hello”.

Gitlab: fixes #2577
Reported-by: phormio