ortho rectIntersect: take input boxes by value
These structs are small enough to pass around by value. And the usage here is
entirely within a single translation unit, so the compiler can see enough to
further optimize. Value semantics are simpler to reason about.

This change is part of a broader goal to reduce the use of `*_at` from the
generic list interface, as an overall simplification. Upcoming changes intend to
alter the relationship between generic list element pointers and their backing
storage, such that retaining a long-lived pointer into the backing storage has
more complicated requirements.