tclpkg allocpoly: swap 'polys_at' for 'polys_back'
This is simpler to read and less error prone.

This change is part of a broader goal to reduce the use of `*_at` from the
generic list interface, as an overall simplification. Upcoming changes intend to
alter the relationship between generic list element pointers and their backing
storage, such that retaining a long-lived pointer into the backing storage has
more complicated requirements.