remove skipping of XPath-based test on CentOS/Rocky
The commit message of 90a811fcb7423298c02c79ca7d7ab254f2a28a09 does not explain,
but I believe this conditional skipping was dealing with Python <3.7. The Python
docs¹ say:

  `[.='text']` Selects all elements whose complete text content, including
               descendants, equals the given text.
               Added in version 3.7.

CentOS testing was removed in 5e1e316900d67cc40dbb35eb3743ecbbca732517. The
Python in use in CI was altered to be a virtual environment in
61aa8f344f6f25e93b5af4b9fc0298c906b856a6, making Rocky 8’s Python 3.11.

¹ https://docs.python.org/3/library/xml.etree.elementtree.html#supported-xpath-syntax