fix mingle ODR violation
Mingle was linking against two copies of common: one directly and one as
common_obj, included in gvc that is then transitively linked through neatogen.
The result of this is that mingle compilation sees two different definitions of
all the variables in lib/common/globals.h. The first one ASan encountered at
runtime was `fdp_parms`:

  ==13717==ERROR: AddressSanitizer: odr-violation (0x595dc2e73c80):
    [1] size=8 'fdp_parms' /home/matthew/graphviz/lib/common/globals.c:33:20
    [2] size=8 'fdp_parms' /home/matthew/graphviz/lib/common/globals.c:33:20
  These globals were registered at these points:
    [1]:
      #0 0x79d60ca37928 in __asan_register_globals ../../../../src/libsanitizer/asan/asan_globals.cpp:341
      #1 0x595dc2d59602 in _sub_I_00099_1 (/tmp/tmp.eYA6q4TsyF/bin/mingle+0x14c602)
      #2 0x79d60aa29eba in call_init ../csu/libc-start.c:145
      #3 0x79d60aa29eba in __libc_start_main_impl ../csu/libc-start.c:379

    [2]:
      #0 0x79d60ca37928 in __asan_register_globals ../../../../src/libsanitizer/asan/asan_globals.cpp:341
      #1 0x79d60beb49a5 in _sub_I_00099_1 (/tmp/tmp.eYA6q4TsyF/bin/../lib/libgvc.so.6+0x4b49a5)
      #2 0x79d60d4ee47d in call_init elf/dl-init.c:70

  ==13717==HINT: if you don't care about these errors you may set ASAN_OPTIONS=detect_odr_violation=0
  SUMMARY: AddressSanitizer: odr-violation: global 'fdp_parms' at /home/matthew/graphviz/lib/common/globals.c:33:20
  ==13717==ABORTING

ODR is only a factor in C++, not in C. So only binaries like Mingle involving
C++ would see this problem. The two definitions of `fdp_parms` (actually the
same one but seen twice) are actually identical, so it seems like this should
not be an ODR violation. However I think one of them being in a shared object
and the other coming transitively from an object library somehow results in
different linkage of the two definitions.