gvc gvrender_resolve_color: use case-insensitive comparison against known colors
This removes a heap allocation as well as opening a path to making font name
pointers const. This is not considered an API break because `knowncolors`
entries were previously required to be lower-cased to match (`canontoken` yields
a lower-case name which will only `strcmp` to 0 against another lower-cased
string). There was no previous use case for non-lower-cased `knowncolors`
entries.