common findStopColor: remove unorthodox allocation scheme for 'clrs'
Memory for both strings `clrs[0]` and `clrs[1]` was being allocated in
`clrs[0]`. `clrs[1]` was then set to an offset into `clrs[0]`. The main
advantage of this seems to have been reducing the allocation and deallocation
calls from 2 to 1. This kind of optimization is no longer relevant in
contemporary environments.

This change zeroes `clrs` in both the initial code of the callee and in the code
of the caller prior to the call. Technically this is unnecessary, but makes this
more robust against future refactorings of either.