neatogen: remove use of globals for priority queue
The use of globals in this code is a barrier to thread-safety, preventing using
multithreading in future.

This change creates the priority queue anew each time through `voronoi`. This is
simpler to reason about for both humans and the compiler. While dropping the
reuse of memory may appear to be a de-optimization, this may actually be an
improvement for the compiler as it can now see the backing memory for the
priority queue does not need to be retained across these calls.

This code still relies on the globals `ymin`, `deltay`, and `sqrt_nsites`.

There is a second, completely distinct priority queue implementation in
lib/neatogen/fPQ.h. It may be worth unifying these in future.

Gitlab: #2558