gvedit: remove generated ui_settings.h; instead generate it during build
The ui_settings.h file is generated by Qt’s User Interface Compiler¹ (UIC). For
unknown reasons, 4df89cce1d9b2e31d29e0bd92c557df6712dae68 checked this into
version control instead of adding a build system rule to generate it. To build
gvedit, the Qt build tools must be available at compilation time,² so avoiding a
dependency on UIC cannot have been the motivation.

This change removes the generated file from the repository and makes it
generated during build now. There are a number of non-obvious characteristics of
this change:

  1. The CMake build was already capable of regenerating this file. I discovered
     this accidentally by deleting cmd/gvedit/ui_settings.h and observing that
     it did not affect the CMake build. With the file checked into the
     repository, CMake previously observed its existence and did not run UIC to
     regenerate it. Following this change, it is now generated during build.

  2. Setting `FORMS` is the documented way of plumbing UI files into a Qt
     project file.³ Yet what the Qt docs seem to fail to mention is that `FORMS`
     is silently ignored if `gui` is not in the defined `QT` list. By not
     defining `QT`, the Autotools build system was producing a fairly neutered
     qMakefile.

  3. `FORMS` is also silently ignored if you make a typo, like adding a .h file
     to this list instead of a .ui file.

  4. The diff between the old checked in file and the new generated
     ui_settings.h is large. Given the checked in copy claims to be from Qt
     4.7.0 which we long ago dropped from support, arguably this seems like a
     fix. I am surprised this has not generated compilation problems before now,
     as the only Qts we support are ≥5.15.2 and ≥6.7.2.

  5. It is important that the `BUILT_SOURCES` line is within a `WITH_QT`-guarded
     block. Without this, compilations that autodetect no Qt or have been passed
     `--with-qt=no` try to run `$(QMAKE)` anyway, which evaluates to `false` and
     then fails the build.

  6. The Qt docs⁴ suggest source-directory-relative paths can be used in project
     files. But if you do this, out-of-source builds get stuck in an infinite
     loop.

Disclaimer: I am neither an Autotools expert nor a Qt expert. It would be fair
to say I am well out of my depth with this change. It is based on reading the Qt
docs and experimentation. The Qt docs alone do not explain anything close to
what is required to make this work. I cannot tell if this is a deficiency in the
Qt docs or a consequence of Graphviz’ Autotools build system making unorthodox
choices in how to interact with Qt.

Gitlab: closes #2579

¹ https://doc.qt.io/qt-6/uic.html
² Here I am distinguishing three environments, (1) `make dist` time, (2)
  compilation time when compilation of the portable source tarball is done, and
  (3) runtime.
³ https://doc.qt.io/qt-6/qmake-variable-reference.html#forms
⁴ https://doc.qt.io/qt-6/qmake-project-files.html